% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_buttons_to_table}
\alias{add_buttons_to_table}
\title{Add buttons to data frame}
\usage{
add_buttons_to_table(
  data,
  id,
  label = icon("arrow-right", class = "fa-regular", lib = "font-awesome"),
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{id}{Character vector. the main id of the buttons.}

\item{label}{Label to use for the button.}

\item{...}{For future expansions. Currently not in use.}
}
\value{
A data frame with a button in each table row.
}
\description{
Small helper function to add Shiny action buttons to a data frame.
}
\examples{
add_buttons_to_table(mtcars[, 1:5], "button_id", "click me") |> 
  datatable_custom()
  
}
