% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup.R
\name{add_shinymanager_auth}
\alias{add_shinymanager_auth}
\title{Add an Authentication Screen}
\usage{
add_shinymanager_auth(app_ui, app_ver, login_note)
}
\arguments{
\item{app_ui}{The \code{app_ui} function for the application.}

\item{app_ver}{See identical param in \code{\link[=run_app]{run_app()}}.}

\item{login_note}{See identical param in \code{\link[=run_app]{run_app()}}.}
}
\description{
Adds an authentication screen via \code{\link[shinymanager:secure-app]{shinymanager::secure_app()}}.
}
