% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup.R
\name{add_tags}
\alias{add_tags}
\title{The 'Add tags' function}
\usage{
add_tags(ui, ...)
}
\arguments{
\item{ui}{placeholder}

\item{...}{placeholder}
}
\description{
The 'Add tags' function
}
\keyword{internal}
