% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addComment.R
\name{addCommentServer}
\alias{addCommentServer}
\title{TODO: comments can't contain "'". Check for other invalid
characters.}
\usage{
addCommentServer(id, metric_abrv, user_name, user_role, pkg_name)
}
\description{
TODO: comments can't contain "'". Check for other invalid
characters.
}
