% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addComment.R
\name{addCommentUI}
\alias{addCommentUI}
\title{The UI for the 'addComment' module}
\usage{
addCommentUI(id)
}
\arguments{
\item{id}{a module id name}
}
\description{
Module to display comments for the specified metric. The comments will update
as the user inserts more comments.
}
