% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup.R
\name{app_theme}
\alias{app_theme}
\title{Application Theme}
\usage{
app_theme()
}
\value{
an bs_theme object of several classes
}
\description{
This sets the Risk Assessment Application theme object using
bslib's bs_theme() function. The app_theme object gets used in run_app.R,
in addition to app_ui.R
}
