% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessmentInfo.R
\name{assessmentInfoUI}
\alias{assessmentInfoUI}
\title{UI for 'Assessment Info' Module}
\usage{
assessmentInfoUI(id)
}
\arguments{
\item{id}{a module id name}
}
\description{
UI for 'Assessment Info' Module
}
