% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auto_font}
\alias{auto_font}
\title{Automatic font re-sizer}
\usage{
auto_font(txt, txt_max = 45, size_min = 0.75, size_max = 1.5, num_bins = NULL)
}
\arguments{
\item{txt}{a string}

\item{txt_max}{an integer to specify a length of text that is considered "to
long" to continue to toggle the font size}

\item{size_min}{an integer specifying the smallest font size you'd like to
see in the output}

\item{size_max}{integer specifying the largest font size you'd like to see in
the output}

\item{num_bins}{when not NULL (the default), accepts an integer that bins a
continuous font size into a categorical one.}
}
\description{
A function that adjusts the number (to be used as font size) that is
proportional to the length of a text string. So the longer the text string,
the smaller the font. Used in MetricBox.R.
}
\examples{
auto_font(txt = "https://github.com/tidyverse/dplyr/issues")
auto_font(txt = "https://github.com/tidyverse/dplyr/issues", num_bins = 2)
auto_font(txt = "https://github.com/tidyverse/dplyr/issues", num_bins = 3)
auto_font(txt = "Yes")
auto_font(txt = "Yes", num_bins = 2)
auto_font(txt = "Yes", num_bins = 3)
}
