% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_build_cards.R
\name{build_db_cards}
\alias{build_db_cards}
\title{The 'Build Database Cards' function}
\usage{
build_db_cards(data)
}
\arguments{
\item{data}{a data.frame}
}
\description{
The 'Build Database Cards' function
}
\keyword{internal}
