% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-csv.R
\docType{data}
\name{community}
\alias{community}
\title{Community metric inclusion basis}
\format{
Data frame with 1 row and 3 variables
}
\usage{
community
}
\description{
A self-generated table that includes community usage metric(s) and why they
were included in the package
}
\keyword{datasets}
