% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityMetrics.R
\name{communityMetricsServer}
\alias{communityMetricsServer}
\title{Community Usage Metrics server logic}
\usage{
communityMetricsServer(id, selected_pkg, community_metrics, user)
}
\description{
Community Usage Metrics server logic
}
