% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_communityMetrics.R
\name{communityMetricsUI}
\alias{communityMetricsUI}
\title{Community Usage Metrics UI function}
\usage{
communityMetricsUI(id)
}
\arguments{
\item{id}{a module id name}
}
\description{
Community Usage Metrics UI function
}
