% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_aboutInfo_utils.R
\name{contrib_card}
\alias{contrib_card}
\title{Contributor Card}
\usage{
contrib_card(role, pic, site, name, org)
}
\arguments{
\item{role}{character, role within the project, usually 'Workstream Lead',
'Core Contributor', or 'Contributor'}

\item{pic}{character, file path to image, 1.3" in width}

\item{site}{character, if they have a personal of GitHub site to reference}

\item{name}{character, first and last name of the contributor}

\item{org}{character, a organization that employs this individual, if
applicable}
}
\description{
Generate a professional card containing role, a photo, name, site and org
}
\examples{
contrib_card(role = "Workstream Lead",
             pic = 'inst/app/www/images/aaron_clark.png',
             site = 'https://github.com/aclark02-arcus',
             name = "Aaron Clark",
             org = "Arcus Biosciences")

}
\seealso{
Other contrib_cards: 
\code{\link{make_contrib_cards}()}
}
\concept{contrib_cards}
\keyword{internal}
