% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_credentials_db}
\alias{create_credentials_db}
\title{Create credentials database}
\usage{
create_credentials_db(db_name = credentials_name)
}
\arguments{
\item{db_name}{a string}
}
\description{
Note: the credentials_name object is assigned in global.R
}
