% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup.R
\name{create_credentials_dev_db}
\alias{create_credentials_dev_db}
\title{Create credentials dev database}
\usage{
create_credentials_dev_db(db_name)
}
\arguments{
\item{db_name}{A string denoting the name of the database}
}
\description{
Create credentials dev database
}
\keyword{internal}
