% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_db}
\alias{create_db}
\title{Create package database}
\usage{
create_db(db_name = database_name)
}
\arguments{
\item{db_name}{a string}
}
\description{
Note: the database_name object is assigned in data-raw/internal-data.R
}
