% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_definition.R
\name{create_fault_tree}
\alias{create_fault_tree}
\title{Tree generation}
\usage{
create_fault_tree(tree_file)
}
\arguments{
\item{tree_file}{is a string with the name of the file}
}
\value{
A Tree object containing the tree strcucture
}
\description{
This function generates a Tree object, with the structure explined later, from a file.
First, it reads a file as a data frame with a specific format.
Then, it computes the tree structure.
}
\details{
The object Tree has the following structure:

The tree structure is defined as a list composed by a set of features of the tree and a set of nodes.
The tree is stored in a file, and a function is provided to read the file and initialize all the data structure. Actually, the features stores in the structure are the following:

Number of Nodes of the tree: tree$n_nodes.
Number of Primary events: tree$n_primary.
Number of Non-primary nodes in the tree, including intermediate nodes and the top event: tree$n_inter.
Identifier of the corner node: tree$corner.
Identifier of the top event node: tree$top.
}
\examples{
Example of use: tree <- create_fault_tree("./tree.conf")

}
