% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseView.R
\name{databaseViewServer}
\alias{databaseViewServer}
\title{Server logic for 'Database View' module}
\usage{
databaseViewServer(id, user, uploaded_pkgs, metric_weights)
}
\description{
Server logic for 'Database View' module
}
