% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{datatable_custom}
\alias{datatable_custom}
\title{Custom datatable}
\usage{
datatable_custom(
  data,
  colnames = c("Package", "Type", "Name", "Version", "Score", "Review Package"),
  hide_names = "name",
  ...
)
}
\arguments{
\item{data}{A data frame as input.}

\item{colnames}{}

\item{hide_names}{Character vector. Whether to hide columns in the data
frame.}

\item{...}{Other options. Currently not in use.}
}
\value{
a DT::datatable object.
}
\description{
Small helper function to create a \code{DT::datatable()} object in a consistent
style.
}
\examples{
datatable_custom(mtcars, colnames = paste0("custom_", names(mtcars)))

}
