% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbSelect}
\alias{dbSelect}
\title{Select data from database}
\usage{
dbSelect(query, db_name = golem::get_golem_options("assessment_db_name"))
}
\arguments{
\item{query}{a sql query as a string}

\item{db_name}{a string}
}
\description{
Select data from database
}
