% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_primary_event_priors.R
\name{derive_primary_event_priors}
\alias{derive_primary_event_priors}
\title{Derive the beta priors on all of the primary event probabilities}
\usage{
derive_primary_event_priors(cornerstone_event_index,
  cornerstone_event_interval, pairwise_matrix)
}
\arguments{
\item{cornerstone_event_index:}{the index of the primary event that is the cornerstone event.
This had to be defined when you defined the tree structure in the file.}

\item{cornerstone_event_interval:}{the range over which the cornerstone event probability is elicited from the expert.
It's mapped to the central 95\% probability interval of a beta distribution.}

\item{pairwise_matrix:}{the matrix of pairwise comparisons. A square matrix whose dimension is the number of primary events.
If a pairwise comparison is not done then the entry is NA.}
}
\value{
A matrix of dimension number of primary events by 2; each row is the
}
\description{
Elicit the beta prior distributions for each primary event probability by the
pairwise comparison approach.
}
\examples{
beta_parameters <- derive_primary_event_priors(cornerstone_event_index = tree$corner,
                                            cornerstone_event_interval  = c(0.01,0.05),
                                            pairwise_matrix             = matrix(c(1,   1/4,  2, 1/2,
                                                                                  4, 1,  2, 4,
                                                                                  1/2, 1/2, 1, 1/5,
                                                                                 (2,   1/4, 5, 1), nrow=4,byrow=TRUE))
Or here is an example of an incomplete comparison matrix that you could also use as the pairwise_matrix
argument in the derive_primary_event_priors function:
pairwise_matrix = matrix(c(1,   1/4,  2,  1/2,
                           4,   1,    NA, NA,
                           1/2, NA,   1,  NA,
                           2,   NA,   NA, 1), nrow=4,byrow=TRUE)
}
