% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_comm_data.R
\name{generate_comm_data}
\alias{generate_comm_data}
\title{generate_comm_data()}
\usage{
generate_comm_data(pkg_name)
}
\arguments{
\item{pkg_name}{A string name of the package.}
}
\value{
A tibble of package usage metrics
}
\description{
generate_comm_data() is a function that extracts usage metrics for a given package.
}
\examples{
metricDF <- generate_comm_data("ggplot2")
metricDF <- generate_comm_data("dplyr")
}
