% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_comm_data}
\alias{get_comm_data}
\title{The 'Get Community Data' function}
\usage{
get_comm_data(pkg_name)
}
\arguments{
\item{pkg_name}{string}
}
\description{
Get all community metric data on a specific package
}
