% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_date_span}
\alias{get_date_span}
\title{The 'Get Date Span' function}
\usage{
get_date_span(start, end = Sys.Date())
}
\arguments{
\item{start}{starting date}

\item{end}{ending date}
}
\description{
Function accepts a start date and optional end date and will
}
