% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_install_pkg_info}
\alias{get_install_pkg_info}
\title{Get the package information from local library}
\usage{
get_install_pkg_info(pkg_name, lib_loc)
}
\arguments{
\item{pkg_name}{string name of the package}

\item{lib_loc}{string name of library directory path}
}
\description{
Get the package information from local library
}
\keyword{internal}
