% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mm_comments}
\alias{get_mm_comments}
\title{The 'Get Maintenance Metrics Comments' function}
\usage{
get_mm_comments(pkg_name)
}
\arguments{
\item{pkg_name}{string}
}
\description{
Retrieves the Maint Metrics comments for a specific package
}
