% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_db.R
\name{get_overall_comments}
\alias{get_overall_comments}
\title{The 'Get Overall Comments' function}
\usage{
get_overall_comments(pkg_name)
}
\arguments{
\item{pkg_name}{string}
}
\description{
Retrieves the overall comments for a specific package
}
