% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_db.R
\name{get_pkg_info}
\alias{get_pkg_info}
\title{The 'Get Package Info' function}
\usage{
get_pkg_info(pkg_name)
}
\arguments{
\item{pkg_name}{string}
}
\description{
Get all general info on a specific package
}
