% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Inference}
\usage{
inference(beta_parameters, data, data_type, MCMC_parameters, tree)
}
\arguments{
\item{beta_parameters:}{A matrix of dimension number of primary events by 2.
It has the beta parameters computed by the derive_primary_event_priors function.}

\item{data:}{A Matrix with number of rows equal to the number of simulated data and number of columns equal to the number of nodes in the tree.
It has the values computed by the Simulate_data function.}

\item{data_type:}{is one of "complete" (all nodes are observed), "top_only" (only the top node is
observed, all others are NA) or "incomplete" (each node is randomly observed with
probability p_obs, otherwise it is NA)}

\item{MCMC_parameters:}{It is a list with the following variables:

n_MCMC:              number of MCMC samples to generate

burn_in:             initial proportion of samples to discard as burn_in

thinning:            save 1 in every thinning samples to use in construction of posterior distributions

logit_p_proposal_sd: the std. dev. of the proposal on the logit p_i}

\item{tree:}{Tree object.}
}
\value{
A matrix that contains samples from the posterior distribution of the event probabilities.
Each row of the matrix is one sample.
Each column is samples for one of the event probabilities, ordered as specified in tree.
}
\description{
This function computes the Posterior distribution of event probabilities
by MCMC and store the results.
It takes the prior beta_parameters (output from derive_primary_event_priors),
the data (might be simulated or real), and
a list of 4 parameters for the MCMC: n_MCMC,  logit_p_proposal_sd, burn_in and thinning.
The output is matrix p_samples.
posterior is a matrix that contains samples from the posterior distribution of the event probabilities.
Each row of the matrix is one sample.
Each column is samples for one of the event probabilities, ordered as specified in tree.
}
\examples{
MCMC_parameters <- list(n_MCMC=10000, burn_in=0.2, thinning=10, logit_p_proposal_sd=0.2)
posterior <- inference(beta_parameters, data, data_type="complete", MCMC_parameters, tree)
}
