% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup.R
\name{initialize_raa}
\alias{initialize_raa}
\title{Initialize the Risk Assessment Application}
\usage{
initialize_raa()
}
\value{
There is no return value. The function is run for its side effects.
}
\description{
This sets up the environment when running the Risk Assessment
Application. It sets the log file, initializes the package database if
missing, and initializes the credentials database if missing.
}
