% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbupload.R
\name{insert_maintenance_metrics_to_db}
\alias{insert_maintenance_metrics_to_db}
\title{The 'Insert MM to DB' Function}
\usage{
insert_maintenance_metrics_to_db(pkg_name)
}
\arguments{
\item{pkg_name}{string name of the package}
}
\description{
Get the maintenance and testing metrics info and upload into DB.
}
