% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_insert_db.R
\name{insert_pkg_info_to_db}
\alias{insert_pkg_info_to_db}
\title{Call function to get and upload info from CRAN/local to db.}
\usage{
insert_pkg_info_to_db(pkg_name)
}
\arguments{
\item{pkg_name}{string name of the package}
}
\description{
Call function to get and upload info from CRAN/local to db.
}
