% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introJS.R
\name{introJSUI}
\alias{introJSUI}
\title{UI for Module to display introJS button and functionality.}
\usage{
introJSUI(id)
}
\arguments{
\item{id}{a module id}
}
\description{
UI for Module to display introJS button and functionality.
}
