% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-csv.R
\docType{data}
\name{maintenance}
\alias{maintenance}
\title{Maintenance metric inclusion basis}
\format{
Data frame with 10 rows and 2 variables
}
\usage{
maintenance
}
\description{
A self-generated table that includes maintenance metric(s) and why they
were included in the package
}
\keyword{datasets}
