% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_maintenanceMetrics.R
\name{maintenanceMetricsServer}
\alias{maintenanceMetricsServer}
\title{Maintenance Metrics module's server logic}
\usage{
maintenanceMetricsServer(id, selected_pkg, maint_metrics, user)
}
\arguments{
\item{id}{a module id name}

\item{selected_pkg}{placeholder}

\item{maint_metrics}{placeholder}

\item{user}{placeholder}
}
\description{
Maintenance Metrics module's server logic
}
