% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_maintenanceMetrics.R
\name{maintenanceMetricsUI}
\alias{maintenanceMetricsUI}
\title{Maintenance Metrics module's UI.}
\usage{
maintenanceMetricsUI(id)
}
\arguments{
\item{id}{a module id name}
}
\description{
Maintenance Metrics module's UI.
}
