% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_aboutInfo_utils.R
\name{make_contrib_cards}
\alias{make_contrib_cards}
\title{Make contributor cards}
\usage{
make_contrib_cards(df = team_info_df)
}
\arguments{
\item{df}{character, role within the project, usually 'Workstream Lead',
'Core Contributor', or 'Contributor'}
}
\description{
Generate group of several 'contributor cards' from a data frame that contains
the following columns:
\itemize{
\item role
\item photo_file
\item site
\item name
-org
}
}
\examples{
make_contrib_cards(team_info_df \%>\% filter(status \%in\% "current"))
make_contrib_cards(team_info_df \%>\% filter(status \%in\% "past"))

}
\seealso{
Other contrib_cards: 
\code{\link{contrib_card}()}
}
\concept{contrib_cards}
\keyword{internal}
