% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_build_cards.R
\name{metric_gauge}
\alias{metric_gauge}
\title{Metric Gauge}
\usage{
metric_gauge(score)
}
\arguments{
\item{score}{a whole number between 0 and 100}
}
\value{
tagList object
}
\description{
An HTML meter element, used for displaying metric scores in the app
}
\keyword{internal}
