% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metricBox.R
\name{metricBoxUI}
\alias{metricBoxUI}
\title{The UI for the 'Metric Box' module}
\usage{
metricBoxUI(id)
}
\arguments{
\item{id}{a module id name}
}
\description{
The UI for the 'Metric Box' module
}
\keyword{internal}
