% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricGrid.R
\name{metricGridServer}
\alias{metricGridServer}
\title{Metric Grid module's server logic}
\usage{
metricGridServer(id, metrics)
}
\arguments{
\item{id}{a module id name}

\item{metrics}{placeholder}
}
\description{
Metric Grid module's server logic
}
