% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_packageDependencies.R
\name{packageDependenciesServer}
\alias{packageDependenciesServer}
\title{Package Dependencies module's server logic}
\usage{
packageDependenciesServer(id, selected_pkg, user, parent, trigger_events)
}
\arguments{
\item{id}{a module id name}

\item{selected_pkg}{placeholder}

\item{user}{placeholder}

\item{parent}{the parent (calling module) session information}

\item{trigger_events}{a reactive values object to trigger actions here or elsewhere}
}
\description{
Package Dependencies module's server logic
}
\keyword{internal}
