% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_packageDependencies.R
\name{packageDependenciesUI}
\alias{packageDependenciesUI}
\title{Package Dependencies module's UI.}
\usage{
packageDependenciesUI(id)
}
\arguments{
\item{id}{a module id name}
}
\description{
Package Dependencies module's UI.
}
\keyword{internal}
