% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reportPreview.R
\name{reportPreviewServer}
\alias{reportPreviewServer}
\title{Server logic for 'Report Preview' module}
\usage{
reportPreviewServer(
  id,
  selected_pkg,
  maint_metrics,
  com_metrics,
  com_metrics_raw,
  mm_comments,
  cm_comments,
  downloads_plot_data,
  dep_metrics,
  loaded2_db,
  user,
  credentials,
  app_version,
  metric_weights
)
}
\arguments{
\item{id}{a module id name}

\item{selected_pkg}{placeholder}

\item{maint_metrics}{placeholder}

\item{com_metrics}{placeholder}

\item{com_metrics_raw}{placeholder}

\item{mm_comments}{placeholder}

\item{cm_comments}{placeholder}

\item{downloads_plot_data}{placeholder}

\item{dep_metrics}{placeholder}

\item{loaded2_db}{placeholder}

\item{user}{placeholder}

\item{app_version}{placeholder}

\item{metric_weights}{placeholder}
}
\description{
Server logic for 'Report Preview' module
}
\keyword{internal}
