% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweightView.R
\name{reweightViewServer}
\alias{reweightViewServer}
\title{Return metric weights. Doing so guarantees that when a report is
downloaded, it would have the latest metric weights.}
\usage{
reweightViewServer(id, user)
}
\description{
Return metric weights. Doing so guarantees that when a report is
downloaded, it would have the latest metric weights.
}
