% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reweightView.R
\name{reweightViewUI}
\alias{reweightViewUI}
\title{UI for the 'Re-weight View' module}
\usage{
reweightViewUI(id)
}
\arguments{
\item{id}{the module id}
}
\description{
UI for the 'Re-weight View' module
}
