% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{package}
\name{riskassessment-package}
\alias{riskassessment}
\alias{riskassessment-package}
\title{The \code{riskassessment} package}
\description{
The Risk Assessment App is an interactive web application serving as a front
end application for the \code{riskmetric} R package. \code{riskmetric} is a framework
to quantify risk by assessing a number of metrics meant to evaluate
development best practices, code documentation, community engagement, and
development sustainability. The app and \code{riskmetric} aim to provide some
context for validation within regulated industries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmaR/riskassessment}
  \item Report bugs at \url{https://github.com/pharmaR/riskassessment/issues}
}

}
\author{
\strong{Maintainer}: Aaron Clark \email{aaron.clark@biogen.com}

Authors:
\itemize{
  \item Robert Krajcik \email{robert.krajcik@biogen.com}
  \item Jeff Thompson \email{jeff.thompson51317@gmail.com}
  \item Marly Gotti \email{marly.cormar@biogen.com}
  \item Maya Gans \email{maya.gans@biogen.com}
  \item Aravind Reddy Kallem
  \item  Fission Labs India Pvt Ltd
}

Other contributors:
\itemize{
  \item  PSI special interest group Application and Implementation of Methodologies in Statistics [copyright holder, funder]
  \item  R Validation Hub [copyright holder, funder]
  \item Munshi Imran Hossain [contributor]
  \item Andrew Borgman \email{andrew.borgman@biogen.com} [contributor]
}

}
\keyword{internal}
