% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  onStart = initialize_raa,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  app_ver = NULL,
  login_note = NULL,
  credentials_db_name = NULL,
  assessment_db_name = NULL,
  decision_categories = c("Low", "Medium", "High"),
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{app_ver}{a "global" variable that is passed to several modules &
reports which details the installed package version when not specified. It
can be overwritten to include a specific version name as a text string.}

\item{login_note}{a text string to display underneath the auth screen's login
button, provided to guide users}

\item{credentials_db_name}{a text string that names the credentials databse.
Please make sure name ends with '.sqlite'. For example: 'cred_db.sqlite'.}

\item{assessment_db_name}{text string that names the credentials databse.
Please make sure name ends with '.sqlite'. For example: 'assess_db.sqlite'.}

\item{decision_categories}{a character vector of the decision categories}

\item{...}{arguments to pass to golem_opts. See \code{?golem::get_golem_options}
for more details.}
}
\value{
a shiny app object
}
\description{
Run the Shiny Application
}
