% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{showComments}
\alias{showComments}
\title{showComments}
\usage{
showComments(pkg_name, comments)
}
\arguments{
\item{pkg_name}{string name of the package}

\item{comments}{data.frame comments table entry}
}
\description{
Displays formatted comments
}
