% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidebar.R
\name{sidebarServer}
\alias{sidebarServer}
\title{Sidebar Server Logic}
\usage{
sidebarServer(id, user, uploaded_pkgs)
}
\arguments{
\item{id}{a module id}

\item{user}{a username}

\item{uploaded_pkgs}{a vector of packages}
}
\description{
Also known as the 'Control Panel', and rightfully so, as it controls
most components of the app, central to it's function
}
\keyword{internal}
