% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebarUI}
\alias{sidebarUI}
\title{Sidebar UI}
\usage{
sidebarUI(id)
}
\arguments{
\item{id}{a module id}
}
\description{
Also known as the 'Control Panel', and rightfully so, as it controls
most components of the app, central to it's function
}
