% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Data}
\usage{
simulate_data(n_simulated_data, tree_definition, data_type, true_primary_p,
  p_obs)
}
\arguments{
\item{tree_definition}{is the fault tree structure to be used, as defined above in the create_fault_tree function.}

\item{true_primary_p}{is an array of the true probabilities of each primary event occurring.
Obviously must be of the same dimension as the number of primary events in tree.}

\item{n_simulated_data:}{number of observations to be generated.}

\item{data_type:}{is one of "complete" (all nodes are observed), "top_only" (only the top node is
observed, all others are NA) or "incomplete" (each node is randomly observed with
probability p_obs, otherwise it is NA)}

\item{p_obs:}{in the case that data_type="incomplete", it is the probability that a node is observed.}
}
\value{
A Matrix with number of rows equal to the number of simulated data and number of columns equal to the number of nodes in the tree.
}
\description{
Functoin that simulate complete data, incomplete data and also approximate the top event probability distribution by simulation.
The data is a matrix type in R with the number of rows equal to the number of observation and the number
       of columns equal to the number of nodes in the tree.  The primary nodes must be listed as
       the first columns and the top event node must be the last column.

    Each element of the matrix is either a 1 (that event was observed to occur), 0 (that event
       was observed not to occur) or NA (that event was not observed)
}
\examples{
data <- simulate_data(n_simulated_data = 5,
                      tree_definition  = tree,
                      data_type        = "complete",
                      true_primary_p   = c(0.02,0.05,0.05,0.1),
                      p_obs            = 0.5)
}
