% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_insert_db.R
\name{update_metric_weight}
\alias{update_metric_weight}
\title{update_metric_weight}
\usage{
update_metric_weight(metric_name, metric_weight)
}
\arguments{
\item{metric_name}{a metric name, as a string}

\item{metric_weight}{a weight, as a string or double}
}
\description{
update_metric_weight
}
