% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-csv.R
\docType{data}
\name{upload_format}
\alias{upload_format}
\title{Upload Format}
\format{
Data frame with 3 rows and 2 variables
}
\usage{
upload_format
}
\description{
A self-generated table that includes an example format that users should
follow when uploading a list of packages to the application
}
\keyword{datasets}
