% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_insert_db.R
\name{upload_pkg_lst}
\alias{upload_pkg_lst}
\title{Upload a Package List}
\usage{
upload_pkg_lst(pkg_lst, assess_db, repos, updateProgress = NULL)
}
\arguments{
\item{pkg_lst}{character vector of packages to upload}

\item{assess_db}{character name (and file path) of the database}

\item{repos}{character vector, the base URL(s) of the repositories to use}

\item{updateProgress}{for internal use only, provides a function to update
progress meter in the application}
}
\value{
A data frame object containing a summary of the upload process
}
\description{
Uploads a list of packages to the database. Designed to be used to set up the
assessment database before deployment of the application.
}
