% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_uploadPackage.R
\name{uploadPackageServer}
\alias{uploadPackageServer}
\title{Server logic for the 'Upload Package' module}
\usage{
uploadPackageServer(id, user, auto_list, credentials, trigger_events, parent)
}
\arguments{
\item{id}{a module id}

\item{user}{a username}

\item{auto_list}{a list of decision automation rules}

\item{trigger_events}{a reactive values object to trigger actions here or elsewhere}

\item{parent}{the parent (calling module) session information}
}
\description{
Server logic for the 'Upload Package' module
}
\keyword{internal}
