% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewComments.R
\name{viewCommentsServer}
\alias{viewCommentsServer}
\title{viewComment module's server logic}
\usage{
viewCommentsServer(id, pkg_name, comments, label = "Current Comments")
}
\arguments{
\item{id}{a module id name}

\item{pkg_name}{placeholder}

\item{comments}{placeholder}

\item{label}{placeholder}
}
\description{
viewComment module's server logic
}
