% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewComments.R
\name{viewCommentsUI}
\alias{viewCommentsUI}
\title{The UI for the 'viewComment' module}
\usage{
viewCommentsUI(id)
}
\arguments{
\item{id}{a module id name}
}
\description{
Module to display comments for the specified metric. The comments will update
as the user inserts more comments.
}
