% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weight_risk_comment}
\alias{weight_risk_comment}
\title{weight_risk_comment}
\usage{
weight_risk_comment(pkg_name)
}
\arguments{
\item{pkg_name}{a package name, as a string}
}
\description{
Used to add a comment on every tab saying how the risk and weights changed,
and that the overall comment & final decision may no longer be applicable.
}
