% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_aboutInfo.R
\name{aboutInfoServer}
\alias{aboutInfoServer}
\title{Server Logic for 'About' Module}
\usage{
aboutInfoServer(id, metric_weights)
}
\arguments{
\item{id}{a module id name}

\item{metric_weights}{object outputted from reweightViewServer()}
}
\description{
Server Logic for 'About' Module
}
\keyword{internal}
