% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addComment.R
\name{addCommentServer}
\alias{addCommentServer}
\title{addComment module's server logic}
\usage{
addCommentServer(id, metric_abrv, user_name, user_role, pkg_name)
}
\arguments{
\item{id}{a module id name}

\item{metric_abrv}{placeholder}

\item{user_name}{placeholder}

\item{user_role}{placeholder}

\item{pkg_name}{string name of the package}
}
\description{
addComment module's server logic
}
