% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup.R
\docType{data}
\name{app_theme}
\alias{app_theme}
\title{Application Theme}
\format{
An object of class \code{bs_bootswatch_lux} (inherits from \code{bs_version_5}, \code{bs_theme}, \code{sass_bundle}) of length 1.
}
\usage{
app_theme
}
\value{
an bs_theme object of several classes
}
\description{
This sets the Risk Assessment Application theme object using
bslib's bs_theme() function. The app_theme object gets used in run_app.R,
in addition to app_ui.R
}
\keyword{datasets}
