% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessmentInfo.R
\name{assessmentInfoServer}
\alias{assessmentInfoServer}
\title{Server Logic for 'Assessment Info' Module}
\usage{
assessmentInfoServer(id, metric_weights)
}
\arguments{
\item{id}{a module id name}

\item{metric_weights}{placeholder}
}
\description{
Server Logic for 'Assessment Info' Module
}
