% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_build_cards.R
\name{build_comm_cards}
\alias{build_comm_cards}
\title{The 'Build Community Cards' function}
\usage{
build_comm_cards(data)
}
\arguments{
\item{data}{a data.frame}
}
\description{
The 'Build Community Cards' function
}
\keyword{internal}
