% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_comm_plotly}
\alias{build_comm_plotly}
\title{The 'Build Community plot' function}
\usage{
build_comm_plotly(data)
}
\arguments{
\item{data}{a data.frame}
}
\description{
The 'Build Community plot' function
}
