% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_dep_cards}
\alias{build_dep_cards}
\title{The 'Build Dependency Cards' function}
\usage{
build_dep_cards(data, loaded, toggled)
}
\arguments{
\item{data}{a data.frame}

\item{loaded}{a vector of package names loaded to db}
}
\description{
The 'Build Dependency Cards' function
}
\keyword{internal}
