% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_communityMetrics.R
\name{communityMetricsServer}
\alias{communityMetricsServer}
\title{Community Usage Metrics server logic}
\usage{
communityMetricsServer(id, selected_pkg, community_metrics, user)
}
\arguments{
\item{id}{a module id name}

\item{selected_pkg}{placeholder}

\item{community_metrics}{placeholder}

\item{user}{placeholder}
}
\description{
Community Usage Metrics server logic
}
