% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_credentials_dev_db}
\alias{create_credentials_dev_db}
\title{Create credentials dev database}
\usage{
create_credentials_dev_db(
  db_name = golem::get_golem_options("credentials_db_name")
)
}
\arguments{
\item{db_name}{a string}
}
\description{
Create credentials dev database
}
