% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup.R
\name{create_db}
\alias{create_db}
\title{Create package database}
\usage{
create_db(db_name)
}
\arguments{
\item{db_name}{A string denoting the name of the database}
}
\description{
Note: the database_name object is assigned by deployment users in R/run_app.R
}
\keyword{internal}
