% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup.R
\name{create_db}
\alias{create_db}
\title{Create package database}
\usage{
create_db(db_name = golem::get_golem_options("assessment_db_name"))
}
\arguments{
\item{db_name}{a string}
}
\description{
Note: the database_name object is assigned by deployment users in R/run_app.R
}
