% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_databaseView.R
\name{databaseViewServer}
\alias{databaseViewServer}
\title{Server logic for 'Database View' module}
\usage{
databaseViewServer(id, user, uploaded_pkgs, metric_weights, changes, parent)
}
\arguments{
\item{id}{a module id name}

\item{user}{a user name}

\item{uploaded_pkgs}{a vector of uploaded package names}

\item{metric_weights}{a reactive data.frame holding metric weights}

\item{changes}{a reactive value integer count}

\item{parent}{the parent (calling module) session information}
}
\description{
Server logic for 'Database View' module
}
\keyword{internal}
