% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbUpdate}
\alias{dbUpdate}
\title{dbUpdate}
\usage{
dbUpdate(command, db_name = database_name)
}
\arguments{
\item{command}{a string}

\item{db_name}{a string}
}
\description{
Deletes, updates or inserts queries.
}
