% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_comm_data}
\alias{generate_comm_data}
\title{Generate Community Usage Data}
\usage{
generate_comm_data(pkg_name)
}
\arguments{
\item{pkg_name}{A string containing the name of a package.}
}
\value{
A tibble of community usage metrics

a data.frame which includes downloads per month for the given pkg
}
\description{
Extracts community usage metrics for a given package.
}
\examples{
if( interactive()) {
ggplot_comm_df <- generate_comm_data("ggplot2")
head(ggplot_comm_df)
}
}
\keyword{reproduce}
