% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cm_comments}
\alias{get_cm_comments}
\title{The 'Get Community Usage Metrics Comments' function}
\usage{
get_cm_comments(pkg_name)
}
\description{
Retrieve the Community Metrics comments for a specific package
}
