% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latest_pkg_info}
\alias{get_latest_pkg_info}
\title{Get the package general information from CRAN/local}
\usage{
get_latest_pkg_info(pkg_name)
}
\arguments{
\item{pkg_name}{string name of the package}
}
\description{
Get the package general information from CRAN/local
}
\keyword{internal}
