% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mm_data}
\alias{get_mm_data}
\title{The 'Get Maintenance Metrics Data' function}
\usage{
get_mm_data(pkg_id)
}
\arguments{
\item{pkg_id}{string}
}
\description{
Pull the maint metrics data for a specific package id, and create
necessary columns for Cards UI
}
