% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_startup.R
\name{initialize_raa}
\alias{initialize_raa}
\title{Initialize \code{riskassessment} Application Settings}
\usage{
initialize_raa(assess_db, cred_db, assess_lib, decision_cat)
}
\arguments{
\item{assess_db}{A string denoting the name of the assessment database.}

\item{cred_db}{A string denoting the name of the credentials database.}

\item{decision_cat}{A character vector denoting the decision categories in ascending order of risk}
}
\value{
There is no return value. The function is run for its side effects.
}
\description{
This sets up the environment when running the \code{riskassessment}
Application. It sets the log file, initializes the package database if
missing, and initializes the credentials database if missing.
}
