% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbupload.R
\name{insert_community_metrics_to_db}
\alias{insert_community_metrics_to_db}
\title{Generate community usage metrics and upload data into DB}
\usage{
insert_community_metrics_to_db(pkg_name)
}
\arguments{
\item{pkg_name}{string name of the package}
}
\description{
Generate community usage metrics and upload data into DB
}
