% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_introJS.R
\name{introJSServer}
\alias{introJSServer}
\title{Server logic for introJS module}
\usage{
introJSServer(id, text, user, approved_roles)
}
\arguments{
\item{id}{a module id}

\item{text}{a data.frame containing info about each step: element names, text
strings guiding users, and a position of where to place the text in
relationship to the element. Please see \code{R/introJSText.R} for data.frames
that populate this argument for the app}

\item{user}{placeholder}
}
\description{
Server logic for introJS module
}
\keyword{internal}
