% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_maintenanceMetrics.R
\name{maintenanceMetricsServer}
\alias{maintenanceMetricsServer}
\title{Maintenance Metrics module's server logic}
\usage{
maintenanceMetricsServer(
  id,
  selected_pkg,
  maint_metrics,
  user,
  approved_roles,
  parent
)
}
\arguments{
\item{id}{a module id name}

\item{selected_pkg}{placeholder}

\item{maint_metrics}{placeholder}

\item{user}{placeholder}

\item{parent}{the parent (calling module) session information}
}
\description{
Maintenance Metrics module's server logic
}
\keyword{internal}
