% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricBox.R
\name{metricBoxServer}
\alias{metricBoxServer}
\title{\code{title}: title.
\code{desc}: description.
\code{value}: metric value.
\code{is_true}: whether the metric is TRUE. If true, then the succ_icon will be
used; if false, then the unsucc_icon will be used.
\code{is_perc}: whether the value is a percentage.
\code{succ_icon}: icon used if is_true.
\code{unsucc_icon}: icon used if not is_true.}
\usage{
metricBoxServer(
  id,
  title,
  desc,
  value,
  is_perc = FALSE,
  is_url = FALSE,
  succ_icon = "check",
  unsucc_icon = "times",
  icon_class = "text-success"
)
}
\description{
\code{title}: title.
\code{desc}: description.
\code{value}: metric value.
\code{is_true}: whether the metric is TRUE. If true, then the succ_icon will be
used; if false, then the unsucc_icon will be used.
\code{is_perc}: whether the value is a percentage.
\code{succ_icon}: icon used if is_true.
\code{unsucc_icon}: icon used if not is_true.
}
