% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricBox.R
\name{metricBoxServer}
\alias{metricBoxServer}
\title{Server logic for the 'Metric Box' module}
\usage{
metricBoxServer(
  id,
  title,
  desc,
  value,
  is_perc = FALSE,
  is_url = FALSE,
  succ_icon = "check",
  unsucc_icon = "times",
  icon_class = "text-success"
)
}
\arguments{
\item{id}{a module id name}

\item{title}{title.}

\item{desc}{description.}

\item{value}{metric value.}

\item{is_perc}{whether the value is a percentage.}

\item{succ_icon}{icon used if is_true.}

\item{unsucc_icon}{icon used if not is_true.}

\item{is_true}{whether the metric is TRUE. If true, then the succ_icon will
be used; if false, then the unsucc_icon will be used.}
}
\description{
Server logic for the 'Metric Box' module
}
