% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_packageDependencies.R
\name{packageDependenciesServer}
\alias{packageDependenciesServer}
\title{Package Dependencies module's server logic}
\usage{
packageDependenciesServer(id, selected_pkg, maint_metrics, user, parent)
}
\arguments{
\item{id}{a module id name}

\item{selected_pkg}{placeholder}

\item{maint_metrics}{placeholder}

\item{user}{placeholder}

\item{parent}{the parent (calling module) session information}
}
\description{
Package Dependencies module's server logic
}
\keyword{internal}
