% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reweightView.R
\name{reweightViewServer}
\alias{reweightViewServer}
\title{Server logic for the 'Re-weight View' module}
\usage{
reweightViewServer(id, user, decision_list, approved_roles, trigger_events)
}
\arguments{
\item{id}{the module id}

\item{user}{the user name}

\item{decision_list}{the list containing the decision automation criteria}

\item{trigger_events}{a reactive values object to trigger actions here or elsewhere}
}
\description{
Server logic for the 'Re-weight View' module
}
\keyword{internal}
