% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweightView.R
\name{reweightViewServer}
\alias{reweightViewServer}
\title{Server logic for the 'Re-weight View' module}
\usage{
reweightViewServer(id, user)
}
\arguments{
\item{id}{the module id}

\item{user}{the user name}
}
\description{
Server logic for the 'Re-weight View' module
}
