% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reweightView.R
\name{reweightViewServer}
\alias{reweightViewServer}
\title{Server logic for the 'Re-weight View' module}
\usage{
reweightViewServer(id, user, decision_list, credentials)
}
\arguments{
\item{id}{the module id}

\item{user}{the user name}

\item{decision_list}{the list containing the decision automation criteria}
}
\description{
Server logic for the 'Re-weight View' module
}
\keyword{internal}
