% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{showHelperMessage}
\alias{showHelperMessage}
\title{Displays a helper message. By default, it informs the user that he should
select a package.}
\usage{
showHelperMessage(message = "Please select a package")
}
\description{
Displays a helper message. By default, it informs the user that he should
select a package.
}
