% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbupload.R
\name{upload_package_to_db}
\alias{upload_package_to_db}
\title{Upload the general info into DB.}
\usage{
upload_package_to_db(
  name,
  version,
  title,
  description,
  authors,
  maintainers,
  license,
  published_on
)
}
\description{
Upload the general info into DB.
}
