% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbupload.R
\name{upload_package_to_db}
\alias{upload_package_to_db}
\title{Upload the general info into DB.}
\usage{
upload_package_to_db(
  name,
  version,
  title,
  description,
  authors,
  maintainers,
  license,
  published_on
)
}
\arguments{
\item{name}{string the package name}

\item{version}{string package version}

\item{title}{string title of the package}

\item{description}{string description of the package}

\item{authors}{string author name(s)}

\item{maintainers}{string names of maintainers}

\item{license}{string type of package license}

\item{published_on}{string char date of publication}
}
\description{
Upload the general info into DB.
}
