% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadPackage.R
\name{uploadPackageServer}
\alias{uploadPackageServer}
\title{Save all the uploaded packages, marking them as 'new', 'not found', or
'duplicate'.}
\usage{
uploadPackageServer(id)
}
\description{
Save all the uploaded packages, marking them as 'new', 'not found', or
'duplicate'.
}
