% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_uploadPackage.R
\name{uploadPackageServer}
\alias{uploadPackageServer}
\title{Server logic for the 'Upload Package' module}
\usage{
uploadPackageServer(id, user)
}
\arguments{
\item{id}{a module id}

\item{user}{a username}
}
\description{
Server logic for the 'Upload Package' module
}
\keyword{internal}
