% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_viewComments.R
\name{viewCommentsServer}
\alias{viewCommentsServer}
\title{viewComment module's server logic}
\usage{
viewCommentsServer(id, pkg_name, comments, label = "Current Comments")
}
\arguments{
\item{id}{the module id name}

\item{pkg_name}{string name of the package}

\item{comments}{data.frame comments table entry}

\item{label}{string default: Current Comments}
}
\description{
viewComment module's server logic
}
